feat: rename use non fungible assets
