+++ packages/plugin-infra/src/web3/index.ts
@@ -78,4 +78,3 @@ export * from './useWeb3Hub'
 export * from './useWeb3State'
 export * from './useWeb3UI'
 export * from './useZeroAddress'
-export * from './useNonFungibleAssets2'
